requirements.txt
cffi==1.17.0
cryptography==43.0.0
customtkinter==5.2.2
darkdetect==0.8.0
mysql-connector-python==9.0.0
packaging==24.1
pillow==10.4.0
pycparser==2.22
python-dotenv==1.0.1
